Allow produce NULL-delimited output (#118)
This commit changes the behavior and declaration of
`inotifytools_sprintf` and `inotifytools_snprintf`. Now it should be
called with new `nstring` struct instead of `char *`. This change allows
to put \0 in buffer for `inotifytools_s(n)printf` because the length of
the buffer is calculated separately without using `strlen`.

It allows you to put \0 and \n in --format option. It can be useful when
parsing inotifywait's output by shell.

It allows to not to put '\n' after user-specified format in the --format
option