clang-format check
Also change .clang-format file, I wanted to use one of the predefined
styles in clang-format, so I didn't have to maintain a .clang-format
file. Our style seemed to not be like any of the pre-defined ones (in
terms of lines of code that match). So I just picked one chromium.