Fix printing of filename when watching a file (#157) (#159)
* Properly terminate escaped CSV strings.

strncpy() does not append a '\0' to the destination string. This results
in the contents of a previous longer string showing at the end of a
shorter string, or garbage if the buffer was not zeroed by the operating
system.

* Fix BSD build error

Return an empty string instead of NULL from csv_escape() to silence
build error:

inotifywait.c: In function 'output_event_csv':
inotifywait.c:144:2: error: '%s' directive argument is null [-Werror=format-overflow=]
  144 |  printf("%s,", csv_escape(inotifytools_event_to_str(event->mask)));

* Skip fanotify OPEN,ISDIR events from self

inotifytools_filename_from_fid() uses open_by_handle_at() to
find path from watch fid that triggers an OPEN,ISDIR event.
Skip this event.

* Test printing of watched filename when watching a file or dir (#157)

inotify prints the watched file as %w and fanotify prints the parent
dir of a watched file as %w.

Verify inotify behavior for backward compatibility until we decide how
to treat this case in fanotify mode.

Verify that both modes print the watched directory.

test

* Fix printing of watched filename when watching a file (#157)

When watching a directory, %f should print the file name and %w should
print the watched directory.

When watching a single file, %f should print an empty string and %w
should print the watched filename.

Commit 0971a97 ("Fanotify support (#134)") changed that for fanotify
because watching a single file or its directory produces the same
fanotify event with non-empty filename.

Revert that change for inotify to fix the regression, but leave it for
fanotify. Update man page to reflect those changes.

Reported-by: auouymous <au@qzx.com>
Signed-off-by: Amir Goldstein <amir73il@gmail.com>